use streams instead of promises as test output
Unlocks a lot of great stuff:
- tests can report many passes and failures per test
- the test suite can continue after failing tests
- streaming test output (woo!)
- decopupled test reportin (woo!)
- trivial to implement TAP output (woo!)

Streams are weird and complicated and have a lot of edge cases that makes them a bit hard to
compose in sticky. Eg., if you `a.pipe(b)`, then when `a` emits `'end'`, `b` is also shut down.
Sticky needs to be able to pipe stream outputs together, but needs to prevent upstreams from
shutting down downstreams.

Anyway stuff seems to roughly work now. Future work on the streaming aspect could mean either
wrapping the streams in something a little more pipe-switching-friendly, or learning more about
streams to do it properly (maybe like `.on('end', a.unpipe(b))` or something???)