Yield one test result per assert
Removes the extra `ok N description` for a failing assert before the `not ok` line, so the reporter outputs just one result
per assert. And doesn't output passing assert lines for asserts that don't pass lol.

Caught up in this is included an actual yaml stringifier for the output details (as per TAP spec), and some general tweaks
to keep everything working together after this.